rename path for minmax datamap
