+++ datamap/examples/src/main/java/org/apache/carbondata/datamap/examples/MinMaxDataWriter.java
