+++ datamap/examples/src/main/java/org/apache/carbondata/datamap/examples/MinMaxIndexBlockDetails.java
